Skewness & Kurtosis(Large & Small Cap Stocks, Fund Indices)
Skewness & kurtosis is another measure of risk in a portfolio. If a fund or stock has positive excess kurtosis and the negative skewness, it implies that large losses are likely.
Computing the skewness and kurtosis are important, as normal distribution assumption is the simplifying assumption that underestimates the magnitude of returns.
There is enough evidence that stock returns and hedge fund index are not normally distributed and we are not looking at a Gaussian distribution.
In order to do justice to the actual return portfolio, we need to extend our analysis beyond mean and variance.